fixed e2e test
