+++ test/e2e/integration/component-library/utils/inputAndVerify.ts
@@ -1,0 +1,5 @@
+export const fillInInputAndVerify = (text: string) => {
+  cy.gotoNavPage('Kort svar');
+  cy.get('#InputPage-Input').type(text);
+  cy.get('[data-testid="summary-single-value-component"]').eq(0).find('span.fds-paragraph').should('have.text', text);
+};
